Moved dom functions from mxml into ChannelML module (only place they're used).
