+++ docs/source/api_index/myokit.mxml.rst
@@ -3,8 +3,6 @@
 ===========
 myokit.mxml
 ===========
-- :meth:`myokit.mxml.dom_child`
-- :meth:`myokit.mxml.dom_next`
 - :meth:`myokit.mxml.html2ascii`
 - :meth:`myokit.mxml.split`
